requirements.txt
paramiko>=2.11.0
scp>=0.14.4
python-dotenv>=1.0.0
